backend/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain[groq]
fastapi[standard]
python-dotenv
pydantic